refactor: rename ImageCopyright to Caption
