+++ src/components/Header/styles.js
@@ -1,9 +1,9 @@
-import styled from "styled-components"
-import { Link } from "gatsby"
+import styled from 'styled-components'
+import { Link } from 'gatsby'
 
-import mediaQuery from "../../utils/mediaQuery"
-import Owl from "../../assets/logo"
-import { navLinkStyle } from "../Nav"
+import mediaQuery from '../../utils/mediaQuery'
+import Owl from '../../assets/logo'
+import { navLinkStyle } from '../Nav'
 
 export const HeaderContainer = styled.header`
   background: ${props => props.theme.darkBlue};
@@ -11,12 +11,12 @@ export const HeaderContainer = styled.header`
   grid-gap: calc(1em + 1vw);
   align-items: center;
   justify-items: center;
-  grid-template-areas: "nav title social search";
+  grid-template-areas: 'nav title social search';
   grid-template-columns: auto 1fr auto auto;
   padding: calc(0.25em + 1vh) calc(1em + 1vw);
   font-weight: bold;
   ${mediaQuery.minNetbook} {
-    grid-template-areas: "title nav social search";
+    grid-template-areas: 'title nav social search';
   }
 `
 
@@ -39,5 +39,4 @@ export const Logo = styled(Owl)`
   border-radius: 50%;
   background: white;
   padding: 0.1em;
-  margin: 0px;
 `
